Fix path separators between operating systems
Fixes an issues where by various path separators could make their way
into the internals of sceptre.

This commit adds a `normalise_path` function that takes a path and
ensures that the correct operating system separator is used when
referencing paths internally and a `sceptreise_path` which will use
`/` regardless of the operating system for use with external services.

This allows users to use either forward or backward slashes in their
project and Sceptre will still work as intended.